ci: add lint+test+coverage+build pipeline; fix ruff defaults and Docker build
Adds .github/workflows/ci.yml with three jobs: lint (ruff + mypy --strict,
GitHub-hosted), test (real Aseprite via self-hosted runner, 100% branch
coverage gate), and docker (build + push, gated on lint). Replaces the old
docker-build.yml, which had no lint/test gate at all.

Fixes found while getting the pipeline green:
- Ruff's now-broader default rule set flagged 16 real issues: preview.py's
  direct subprocess/file calls inside async def are wrapped in
  asyncio.to_thread instead of blocking the event loop; narrowed
  except Exception to concrete types where the real exception set is known,
  kept it (with a noqa + rationale) at the three MCP tool-boundary catches
  where an unexpected error must report back to the caller rather than crash
  the server; dropped a redundant int(round(...)) cast; switched a couple of
  bare open() calls to context managers.
- Dockerfile's steamcmd install never actually worked: it's in Ubuntu's
  multiverse component and needs the i386 arch plus non-interactive EULA
  acceptance, none of which were configured. Fixed and build-verified.
- .dockerignore didn't exclude .env, so a future real Steam credential file
  would have been baked into every image layer; excluded it (and cache dirs
  that were leaking into the build context).
- ci.yml's self-hosted test job now sets ASEPRITE_PATH explicitly rather
  than relying on the gitignored, checkout-absent .env file.

+++ aseprite_mcp/core/fonts.py
@@ -47,8 +47,8 @@ from __future__ import annotations
 
 import json
 import os
-from dataclasses import dataclass, field
 from collections.abc import Iterator
+from dataclasses import dataclass, field
 from typing import Any
 
 from PIL import Image, ImageDraw, ImageFont
@@ -306,7 +306,7 @@ class TrueTypeFont:
     ) -> tuple[set[Point], int]:
         """Rasterise `text`; x runs from the pen origin, y from the baseline."""
         ascent, descent = font.getmetrics()
-        advance = int(round(font.getlength(text)))
+        advance = round(font.getlength(text))
         box = font.getbbox(text)
         # Glyphs can overhang the advance on either side (italics, accents,
         # negative bearings), so leave room around the drawn string.
